Abstract Notification class instead of interface
The abstract class can extend the laravel Notification class, provide
the default implementation of the `via` method which won't change in
child classes and define the `toTwitter` method that needs to be
added in child classes.
Also change the argument in the TwitterChannel::send method to
TwitterNotification as this will be included in a new major version.