fix immediately discarded twitter version
The `changeTwitterSettingsIfNeeded` method creates a new instance of the TwitterOAuth class. 

The default API version in TwitterOAuth is 2. 

Setting the api version *before* changing the twitter settings creates an opportunity for the api version to be **immediately** overwritten if the settings need to be changed.

Flipping these lines is all that is necessary to remediate the error.